Add capybara gem
Capybara uses a "headless browser" (eg a browser without a UI) to
actually visit our application and test it (click links, make sure the
proper text shows up, etc.)

This enables us to write "feature" specs, which are known in the rest of
the industry as integration tests. They will live in the spec/features
directory.