Add formatting to home page
This evaluates to:

<h1>Hello World of Railing Stocks</h1>

<h2>The current time is [time].</h2>

When you use a % symbol to represent an html tag, you can either place
single-line content on the same line as the symbol, or you can use
multi-line content under the symbol, as long as you indent your content.
I will give an example of this in my next commit.