Remove 'bundled with' message
If you see this message, update your bundler (I think `gem update
bundler` will do the trick) and hopefully it will go away. Let's try not
to commit the Gemfile.lock if this message appears in it (I know, I'm
probably the one who did it in the first place).